Add DeviceModule MixIn class
