+++ tests/model/test_device_module.py
@@ -1,0 +1,6 @@
+from document_segmentation.model.device_module import DeviceModule
+
+
+class TestDeviceModule:
+    def test_get_device(self):
+        assert DeviceModule.get_device() in {"cuda", "mps", "cpu"}
